// frontend/src/api/generated/models/learningRoadmapDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SkillPilot API
 * OpenAPI spec version: v1
 */

export interface LearningRoadmapDto {
  id?: string;
  /** @nullable */
  skillResultId?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  week?: number;
  completed?: boolean;
}